Migrate UrlUtils to TypeScript
